Improved CDC initialization: fixed sleep -> waiting for service to bee healthy

+++ IO.Eventuate.Tram.Tests/test.sh
@@ -3,27 +3,52 @@
 set -e
 set -x
 
+KEEP_RUNNING=
+USE_EXISTING_CONTAINERS=
+
+while [ ! -z "$*" ] ; do
+  case $1 in
+    "--keep-running" )
+      KEEP_RUNNING=yes
+      ;;
+    "--use-existing-containers" )
+      USE_EXISTING_CONTAINERS=yes
+      ;;
+    "--help" )
+      echo ./test.sh --keep-running --use-existing-containers
+      exit 0
+      ;;
+  esac
+  shift
+done
+
 if dotnet nuget list source | grep -q 'https://api.bintray.com/nuget/eventuateio-oss/eventuateio-dotnet-snapshots'; then
   echo "Package already exists"
 else
-  echo "Add package" 
+  echo "Add package"
   dotnet nuget add source https://api.bintray.com/nuget/eventuateio-oss/eventuateio-dotnet-snapshots
 fi
 
 dotnet build -c release
 
-docker-compose down
+if [ -z "$USE_EXISTING_CONTAINERS" ] ; then
+  docker-compose down
+fi
+
 docker-compose up -d mssql
 
 docker-compose run --rm wait-for-db
 
-docker-compose run --rm dbsetup
+if [ -z "$USE_EXISTING_CONTAINERS" ] ; then
+  docker-compose run --rm dbsetup
+fi
+  
 docker-compose up -d zookeeper
 docker-compose up -d kafka
 docker-compose up -d cdcservice
 
 # Wait for docker containers to start up
-sleep 60s
+./wait-for-services.sh
 
 # Print docker status
 docker ps
